% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatedSwimmerPlot.R
\name{animatedSwimmerPlot}
\alias{animatedSwimmerPlot}
\title{<Add Title>}
\usage{
animatedSwimmerPlot(
  data,
  settings = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
